@@ -5,7 +5,7 @@
   "license": "Apache-2.0",
   "author": "Google Inc.",
   "engines": {
-    "node": ">=8.10.0"
+    "node": ">=10"
   },
   "repository": "googleapis/nodejs-speech",
   "main": "./build/src/index.js",
@@ -44,7 +44,7 @@
   },
   "dependencies": {
     "@google-cloud/common": "^2.0.0",
-    "google-gax": "^1.11.1",
+    "google-gax": "^2.0.1",
     "protobufjs": "^6.8.6",
     "pumpify": "^2.0.0",
     "stream-events": "^1.0.4",
@@ -53,14 +53,14 @@
   "devDependencies": {
     "@types/mocha": "^7.0.0",
     "@types/node": "^12.0.0",
-    "@types/sinon": "^7.5.1",
+    "@types/sinon": "^7.5.2",
     "c8": "^7.0.0",
     "codecov": "^3.0.2",
     "eslint": "^6.0.0",
     "eslint-config-prettier": "^6.0.0",
     "eslint-plugin-node": "^11.0.0",
     "eslint-plugin-prettier": "^3.0.0",
-    "gts": "^1.1.2",
+    "gts": "2.0.0-alpha.9",
     "jsdoc": "^3.5.5",
     "jsdoc-fresh": "^1.0.1",
     "jsdoc-region-tag": "^1.0.2",
@@ -69,9 +69,9 @@
     "null-loader": "^3.0.0",
     "pack-n-play": "^1.0.0-2",
     "prettier": "^1.13.5",
-    "sinon": "^9.0.0",
+    "sinon": "^9.0.1",
     "ts-loader": "^6.2.1",
-    "typescript": "3.6.4",
+    "typescript": "^3.8.3",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10"
   }
