@@ -27,17 +27,17 @@ const storage = new Storage();
 const bucketName = `nodejs-docs-samples-test-${uuid.v4()}`;
 const cmd = 'node recognize.js';
 const resourcePath = path.join(__dirname, '..', 'resources');
-const filename = `audio.raw`;
-const filename1 = `Google_Gnome.wav`;
-const filename2 = `commercial_mono.wav`;
-const filename3 = `commercial_stereo.wav`;
+const filename = 'audio.raw';
+const filename1 = 'Google_Gnome.wav';
+const filename2 = 'commercial_mono.wav';
+const filename3 = 'commercial_stereo.wav';
 const filepath = path.join(resourcePath, filename);
 const filepath1 = path.join(resourcePath, filename1);
 const filepath2 = path.join(resourcePath, filename2);
 const filepath3 = path.join(resourcePath, filename3);
 const text = 'how old is the Brooklyn Bridge';
 const text1 = 'the weather outside is sunny';
-const text2 = `Terrific. It's on the way.`;
+const text2 = "Terrific. It's on the way.";
 const text3 = 'Chrome';
 
 describe('Recognize', () => {
